fixed staff selecting/resizing and finally made delete button work in all cases
